requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.1
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2024.2.2
charset-normalizer==2.1.0
click==8.0.3
coloredlogs==15.0.1
fastapi==0.72.0
filelock==3.7.1
filetype==1.0.9
Flask==2.2.2
flatbuffers==1.12
frozenlist==1.3.1
gast==0.4.0
gdown==4.5.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
humanfriendly==10.0
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.3
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
mpmath==1.3.0
multidict==6.0.2
networkx==2.8.5
numba==0.56.0
numpy==1.22.3
oauthlib==3.2.0
onnxruntime==1.18.0
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
platformdirs==4.2.2
pooch==1.8.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
PyMatting==1.1.7
pyparsing==3.0.9
pyrsistent==0.20.0
PySocks==1.7.1
python-multipart==0.0.5
PyWavelets==1.3.0
rembg==2.0.56
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.1
scikit-learn==1.1.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.17.1
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.8
tqdm==4.62.3
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.17.0
watchdog==2.1.7
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1